Introduce `CreateProject` section with improved structure and components.
- Move `CreateProject`-related files into a dedicated subnamespace for better modularization.
- Add `CreateProjectView`, `CreateProjectStage`, and `CreateProjectStageDesign` classes.
- Enhance `DetailsView` and `ProfileView` with updated layouts, bindings, and additional styling.
- Add `ICreateProjectStage` interface and adjust `CreateProjectViewModel` implementation.
- Refactor UI components with improved binding and namespace alignment.
