+++ src/Angor/Avalonia/AngorApp/Sections/Founder/CreateProject/CreateProjectView.axaml
@@ -1,0 +1,15 @@
+<UserControl xmlns="https://github.com/avaloniaui"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+             xmlns:cp="clr-namespace:AngorApp.Sections.Founder.CreateProject"
+             mc:Ignorable="d" d:DesignWidth="800" 
+             x:Class="AngorApp.Sections.Founder.CreateProject.CreateProjectView">
+    <ScrollViewer>
+        <StackPanel Spacing="10">
+            <cp:ProfileView />
+            <cp:DetailsView />
+        </StackPanel>
+    </ScrollViewer>
+
+</UserControl>
