+++ src/Angor/Avalonia/AngorApp/Sections/Founder/CreateProject/CreateProjectView.axaml.cs
@@ -1,0 +1,13 @@
+using Avalonia;
+using Avalonia.Controls;
+using Avalonia.Markup.Xaml;
+
+namespace AngorApp.Sections.Founder.CreateProject;
+
+public partial class CreateProjectView : UserControl
+{
+    public CreateProjectView()
+    {
+        InitializeComponent();
+    }
+}
