+++ src/Angor/Avalonia/AngorApp/Sections/Founder/CreateProject/ICreateProjectStage.cs
@@ -1,0 +1,10 @@
+using Zafiro.UI.Commands;
+
+namespace AngorApp.Sections.Founder.CreateProject;
+
+public interface ICreateProjectStage
+{
+    public double Percent { get; set; }
+    public DateTimeOffset? StartDate { get; set; }
+    public IEnhancedCommand Remove { get; }
+}
