+++ src/Angor/Avalonia/AngorApp/Sections/Founder/CreateProject/ICreateProjectViewModel.cs
@@ -1,0 +1,13 @@
+using Zafiro.UI.Commands;
+
+namespace AngorApp.Sections.Founder.CreateProject;
+
+public interface ICreateProjectViewModel
+{
+    public IEnumerable<ICreateProjectStage> Stages { get; }
+    public DateTime StartDate { get; }
+    public DateTime? EndDate { get; set; }
+    public IEnhancedCommand AddStage { get; }
+    public int? PenaltyDays { get; set; }
+    public DateTime? ExpiryDate { get; }
+}
