+++ src/Angor/Avalonia/AngorApp/Sections/Founder/CreateProject/ProfileView.axaml
@@ -2,23 +2,37 @@
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-             mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="450"
+             xmlns:ui="clr-namespace:Zafiro.UI;assembly=Zafiro.UI"
+             xmlns:controls="clr-namespace:AngorApp.UI.Controls"
+             xmlns:collections="clr-namespace:System.Collections;assembly=System.Runtime"
+             mc:Ignorable="d" d:DesignWidth="700" 
              x:Class="AngorApp.Sections.Founder.CreateProject.ProfileView">
-    <StackPanel Spacing="20">
-        <HeaderedContainer Header="Project name">
-            <TextBox Watermark="Enter the name of the project" />
-        </HeaderedContainer>
-        <HeaderedContainer Header="About">
-            <TextBox Watermark="Provide a brief description of the project" />
-        </HeaderedContainer>
-        <HeaderedContainer Header="Project website">
-            <TextBox Watermark="Enter the name of the project" />
-        </HeaderedContainer>
-        <HeaderedContainer Header="Banner Image URL (Recommended size: 820×312px)">
-            <TextBox Watermark="Enter the URL for the project banner image" />
-        </HeaderedContainer>
-        <HeaderedContainer Header="Avatar Image Url (Recommended size: 170×170px)">
-            <TextBox Watermark="Enter the URL for the project logo or profile picture" />
-        </HeaderedContainer>
-    </StackPanel>
+    <Card Header="Project Profile">
+        <Card.HeaderStartContent>
+            <ui:Icon Source="fa-briefcase" />
+        </Card.HeaderStartContent>
+        <StackPanel Spacing="10">
+            <controls:CardGrid ColumnSpacing="10" RowSpacing="10">
+                <controls:CardGrid.ItemsSource>
+                    <collections:ArrayList>
+                        <HeaderedContainer Header="Project name">
+                            <TextBox Watermark="Enter the name of the project" />
+                        </HeaderedContainer>
+                        <HeaderedContainer Header="About">
+                            <TextBox Watermark="Provide a brief description of the project" />
+                        </HeaderedContainer>
+                    </collections:ArrayList>
+                </controls:CardGrid.ItemsSource>
+            </controls:CardGrid>
+            <HeaderedContainer Header="Project website">
+                <TextBox Watermark="Enter the name of the project" />
+            </HeaderedContainer>
+            <HeaderedContainer Header="Banner Image URL (Recommended size: 820×312px)">
+                <TextBox Watermark="Enter the URL for the project banner image" />
+            </HeaderedContainer>
+            <HeaderedContainer Header="Avatar Image Url (Recommended size: 170×170px)">
+                <TextBox Watermark="Enter the URL for the project logo or profile picture" />
+            </HeaderedContainer>
+        </StackPanel>
+    </Card>
 </UserControl>
